Remove unknown `type` in configuration array
The snippet passes a `type => ‘web’` in the configuration array - but there is no indication of the meaning of this.
A search in the PageDoktypeRegistry was unsuccessful. The entry is not evaluated here. Nothing could be found in the rest of the core either. And even when asked via Mastodon, there was no indication of the meaning/necessity.