Update CreateModuleWithExtbase.rst (#5138) (#5140)
Extend the code example how to create links to Extbase controller actions. It is not obvious based on the current documentation, that the array-key of the new module-configuration must be used, plus passing the controller and action name as arguments.

The issue was solved in Slack (thanks Garvin!) and is now added here for future use.

Co-authored-by: Hanns <1685510+moongazer@users.noreply.github.com>